+++ include/graphics_data.h
@@ -3,13 +3,13 @@
 
 #include "prelude.h"
 
-extern u8 const Img_Unk_02017374[]; // lz
-extern u8 const Tm_Unk_02017AA8[];  // lz
-extern u8 const Tm_Unk_02017908[];  // lz
+extern u32 const Img_Unk_02017374[]; // lz
+extern u32 const Tm_Unk_02017AA8[];  // lz
+extern u32 const Tm_Unk_02017908[];  // lz
 extern u16 const Pal_Unk_02017C74[];
 
 #if defined(VER_FINAL)
-extern u8 const Tm_Unk_020179D8[]; // lz
+extern u32 const Tm_Unk_020179D8[]; // lz
 #endif
 
 #endif // UNKNOWNS_H
